+++ model/table.go
@@ -99,7 +99,7 @@ func (t Table) SQL() string {
 				if col.Default != nil {
 					q += " GENERATED ALWAYS AS (" + *col.Default + ") STORED"
 				}
-			} else if col.Default != nil {
+			} else if col.Default != nil && col.SerialSequence == nil {
 				q += " DEFAULT " + *col.Default
 			}
 			if col.NotNull && !col.Identity.IsIdentityColumn() {